Add an option on the connection object to set the encoding of query results, since we only allow a connection to be used for a single host.  This also required changing the initializers for other objects to accept an encoding parameters.
Further checking on yaz encoding issues revealed that we need to explicitly
convert from marc-8 for "raw" keys in records, so we now perform this
conversion for USMARC and UKMARC syntax types.  Needs reliability testing,
but fixes issues with the Library of Congress server and opac.nebis.ch.

Records from COPAC will now be translated incorrectly unless a fallback
encoding of iso-8859-1 is set, or the connection encoding is explicitly
set to iso-8859-1.