Database find & replace redesign.
Place the Replace All options and Find options in boxes, so the title gioves more info about their function. 

Replace replace all scope popup by a matrix, so it is more obvious that the scope can be the shown items as opposed to the whole database.

Add a Wrap Around option. 

 Fix enabling of search scope popup.